Test castro

+++ EquiposTecnicosSN.Web/Controllers/FabricantesController.cs
@@ -1,9 +1,11 @@
 ﻿using System.Data.Entity;
-using System.Threading.Tasks;
 using System.Net;
 using System.Web.Mvc;
 using EquiposTecnicosSN.Web.DataContexts;
 using EquiposTecnicosSN.Entities.Equipos.Info;
+using System.Linq;
+using PagedList;
+using System.Collections.Generic;
 
 namespace EquiposTecnicosSN.Web.Controllers
 {
@@ -13,19 +15,74 @@ namespace EquiposTecnicosSN.Web.Controllers
         private EquiposDbContext db = new EquiposDbContext();
 
         // GET: Fabricantes
-        public async Task<ActionResult> Index()
+        public ActionResult Index(string searchNombre = null, string searchMarca = null, string searchModelo = null, int page = 1)
         {
-            return View(await db.Fabricantes.ToListAsync());
+            int? fabricanteId = null;
+            List<int> fabricantesId = new List<int>();
+            bool checkFabricanteId = false;
+            bool checkFabricantesIds = false;
+            if (searchMarca != null && searchMarca != "")
+            {
+                var marcas = db.Marcas.Where(m => m.Nombre.Contains(searchMarca)).ToList();
+                if  (marcas.Count == 1)
+                {
+                    checkFabricanteId = true;
+                    fabricanteId = marcas.First().FabricanteId;
+                } else if (marcas.Count > 1)
+                {
+                    checkFabricantesIds = true;
+                    fabricantesId = marcas.Select(m => m.FabricanteId).ToList();
+                } else if (marcas.Count == 0)
+                {
+                    checkFabricanteId = true;
+                    checkFabricantesIds = true;
+                }
+
+            }
+
+            if (searchModelo != null && searchModelo != "")
+            {
+                var modelos = db.Modelos.Where(mod => mod.Nombre.Contains(searchModelo)).ToList();
+                if (modelos.Count == 1)
+                {
+                    checkFabricanteId = true;
+                    fabricanteId = modelos.First().Marca.FabricanteId;
+                } else if (modelos.Count > 1)
+                {
+                    checkFabricantesIds = true;
+                    fabricantesId = modelos.Select(m => m.Marca.FabricanteId).ToList();
+                }
+                else if (modelos.Count == 0)
+                {
+                    checkFabricanteId = true;
+                    checkFabricantesIds = true;
+                }
+            }
+
+            var listPage = db.Fabricantes
+                .Where(f => (searchNombre == null || searchNombre == "") || f.Nombre.Contains(searchNombre))
+                .Where(f => (!checkFabricanteId && fabricanteId == null) || f.FabricanteId == fabricanteId)
+                .Where(f => (!checkFabricantesIds && fabricantesId.Count == 0) || fabricantesId.Contains(f.FabricanteId))
+                .OrderBy(f => f.Nombre)
+                .ToPagedList(page, 10);
+
+            if (Request.IsAjaxRequest())
+            {
+                return PartialView("_FabricantesList", listPage);
+            }
+
+            return View(listPage);
+
         }
 
         // GET: Fabricantes/Details/5
-        public async Task<ActionResult> Details(int? id)
+        public ActionResult Details(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Fabricante fabricante = await db.Fabricantes.FindAsync(id);
+            Fabricante fabricante = db.Fabricantes.Find(id);
             if (fabricante == null)
             {
                 return HttpNotFound();
@@ -36,7 +93,7 @@ namespace EquiposTecnicosSN.Web.Controllers
         // GET: Fabricantes/Create
         public ActionResult Create()
         {
-            return View();
+            return View(new Fabricante());
         }
 
         // POST: Fabricantes/Create
@@ -44,12 +101,12 @@ namespace EquiposTecnicosSN.Web.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Create([Bind(Include = "FabricanteId,Nombre")] Fabricante fabricante)
+        public ActionResult Create(Fabricante fabricante)
         {
             if (ModelState.IsValid)
             {
                 db.Fabricantes.Add(fabricante);
-                await db.SaveChangesAsync();
+                db.SaveChanges();
                 return RedirectToAction("Index");
             }
 
@@ -57,13 +114,13 @@ namespace EquiposTecnicosSN.Web.Controllers
         }
 
         // GET: Fabricantes/Edit/5
-        public async Task<ActionResult> Edit(int? id)
+        public ActionResult Edit(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Fabricante fabricante = await db.Fabricantes.FindAsync(id);
+            Fabricante fabricante = db.Fabricantes.Find(id);
             if (fabricante == null)
             {
                 return HttpNotFound();
@@ -72,29 +129,27 @@ namespace EquiposTecnicosSN.Web.Controllers
         }
 
         // POST: Fabricantes/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Edit([Bind(Include = "FabricanteId,Nombre")] Fabricante fabricante)
+        public ActionResult Edit(Fabricante fabricante)
         {
             if (ModelState.IsValid)
             {
                 db.Entry(fabricante).State = EntityState.Modified;
-                await db.SaveChangesAsync();
+                db.SaveChanges();
                 return RedirectToAction("Index");
             }
             return View(fabricante);
         }
 
         // GET: Fabricantes/Delete/5
-        public async Task<ActionResult> Delete(int? id)
+        public ActionResult Delete(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Fabricante fabricante = await db.Fabricantes.FindAsync(id);
+            Fabricante fabricante = db.Fabricantes.Find(id);
             if (fabricante == null)
             {
                 return HttpNotFound();
@@ -105,12 +160,36 @@ namespace EquiposTecnicosSN.Web.Controllers
         // POST: Fabricantes/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> DeleteConfirmed(int id)
+        public ActionResult DeleteConfirmed(int id)
         {
-            Fabricante fabricante = await db.Fabricantes.FindAsync(id);
-            db.Fabricantes.Remove(fabricante);
-            await db.SaveChangesAsync();
-            return RedirectToAction("Index");
+            var equiposCount = db.Equipos
+                .Where(e => e.InformacionHardware.FabricanteId == id)
+                .Count();
+
+            var marcasCount = db.Marcas
+                .Where(m => m.FabricanteId == id)
+                .Count();
+
+            var fabricante = db.Fabricantes.Find(id);
+
+            if (equiposCount == 0 && marcasCount == 0)
+            {                
+                db.Fabricantes.Remove(fabricante);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+
+            if (equiposCount > 0)
+            {
+                ModelState.AddModelError("", "El Fabricante no puede eliminarse ya que posee equipos asociados.");
+            }
+
+            if (marcasCount > 0)
+            {
+                ModelState.AddModelError("", "El Fabricante no puede eliminarse ya que posee Marcas asociados.");
+            }
+
+            return View(fabricante);
         }
 
         protected override void Dispose(bool disposing)
